==> Oakland/01_Data Extraction I.ipynb <==
import csv

# DataSet

# 2 week Extension (Aug)
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('00Extension.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(17,32):
        with open('C:/Oakland/00Aug/d04_text_station_5min_2017_08_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Sept 2017
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('01Sept.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,31):
        with open('C:/Oakland/01Sept/d04_text_station_5min_2017_09_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Oct 2017
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('02Oct.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,32):
        with open('C:/Oakland/02Oct/d04_text_station_5min_2017_10_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Nov 2017
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('03Nov.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,31):
        with open('C:/Oakland/03Nov/d04_text_station_5min_2017_11_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Dec 2017
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('04Dec.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,32):
        with open('C:/Oakland/04Dec/d04_text_station_5min_2017_12_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Jan 2018
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('05Jan.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,32):
        with open('C:/Oakland/05Jan/d04_text_station_5min_2018_01_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Feb 2018
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('06Feb.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,29):
        with open('C:/Oakland/06Feb/d04_text_station_5min_2018_02_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)

# Mar 2018
stations = ['408588','401190','408589','401144','408590','401615','408591','401413','401896','401897','400642']

with open('07Mar.csv', 'w') as new_file:
    csv_writer = csv.writer(new_file)
    
    for i in range(1,5):
        with open('C:/Oakland/07Mar/d04_text_station_5min_2018_03_{}.txt'.
                  format(str(i).zfill(2)), 'r') as csv_file:
            csv_reader = csv.reader(csv_file)
    
            for line in csv_reader:
                if line[1] in stations:
                    csv_writer.writerow(line)
