==> Oakland/02_Data Extraction II.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# DataSet

# Read saved csv files
Ext = pd.read_csv('00Extension.csv',header=None)
Sep = pd.read_csv('01Sept.csv',header=None)
Oct = pd.read_csv('02Oct.csv',header=None)
Nov = pd.read_csv('03Nov.csv',header=None)
Dec = pd.read_csv('04Dec.csv',header=None)
Jan = pd.read_csv('05Jan.csv',header=None)
Feb = pd.read_csv('06Feb.csv',header=None)
Mar = pd.read_csv('07Mar.csv',header=None)

# Trim the data (first 17 columns)
Ext = Ext.iloc[:,0:17]
Sep = Sep.iloc[:,0:17]
Oct = Oct.iloc[:,0:17]
Nov = Nov.iloc[:,0:17]
Dec = Dec.iloc[:,0:17]
Jan = Jan.iloc[:,0:17]
Feb = Feb.iloc[:,0:17]
Mar = Mar.iloc[:,0:17]

# Define dataframe header
columns = ['Timestamp','Station','District','Freeway','DoT','Lane Type','Station Length','Samples', '% Obs',
          'Total flow','Avg Occ','Ave Speed','Lane N Samples','Lane N Flow','Lane N Avg Occ','Lane N Avg Speed','Lane N obs']

# Input the header defined above
Ext.columns = columns
Sep.columns = columns
Oct.columns = columns
Nov.columns = columns
Dec.columns = columns
Jan.columns = columns
Feb.columns = columns
Mar.columns = columns

Oct.head()

Oct.tail()

df = pd.concat([Ext,Sep,Oct,Nov,Dec,Jan,Feb,Mar],ignore_index=True)

len(df)

# Pass a list of column names (the important variables)
df = df[['Timestamp','Station','Total flow']]

df['Timestamp'] = pd.to_datetime(df['Timestamp'], infer_datetime_format=True)

pd.to_datetime(df['Timestamp'], infer_datetime_format=True)

df

# Replace Station ID
mymap = {
            408588  : 'f1',
            401190  : 'f2',
            408589  : 'f3',
            401144  : 'f4',
            408590  : 'f5',
            401615  : 'f6',
            408591  : 'f7',
            401413  : 'f8',
            401896  : 'f9',
            401897  : 'f10',
            400642  : 'f11'
        }
df['Station'] = df['Station'].map(mymap)

df['Station'].value_counts()

df['Timestamp'].nunique()

# Re-arrange the Dataframe to spatio-temporal format
df = df.pivot(index='Timestamp',columns='Station',values='Total flow')

df.keys()

df

# Drop f1; all nan values
df.drop(columns=['f1'],inplace=True)

# Re-arrange spatial columns, so, the data is sequential
df = df[['f2','f3','f4','f5','f6','f7','f8','f9','f10','f11']]
cols = ['f1','f2','f3','f4','f5','f6','f7','f8','f9','f10']
df.columns = cols
df

df.to_csv('000TrainVal.csv')
