==> Oakland/03c_DPP_Data Scaling.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Read saved csv files
TrainVal     = pd.read_csv('0TrainVal.csv',index_col=0,parse_dates=True)

TrainVal

# Split to Training and Test Data

len(TrainVal)

final_2weeks = 4032
train_end = len(TrainVal) - final_2weeks    

train_end

train = TrainVal.iloc[:train_end]      # From 1st element to 53568
test = TrainVal.iloc[train_end:]       # From 53568 to end

train.tail()

test.head()

train.index

# Feature Scaling

# The scaler - for min-max scaling
from sklearn.preprocessing import MinMaxScaler

The MinMaxScaler from scikit-learn subtracts the minimum value and re-scales the the value range, rendering the variable values between 0 and 1. 

# set up the scaler
scaler = MinMaxScaler()

# fit the scaler to the train set, it will learn the parameters
scaler.fit(train)

# transform train and test sets
train_scaled = scaler.transform(train)
test_scaled = scaler.transform(test)

# the scaler stores the maximum values of the features, learned from train set
scaler.data_max_

# the scaler stores the minimum values of the features, learned from train set
scaler.data_min_

l = ['f1','f2','f3','f4','f5','f6']

train_scaled = pd.DataFrame(train_scaled, columns=l)
test_scaled = pd.DataFrame(test_scaled, columns=l)

train_scaled.set_index(train.index,  inplace=True);
test_scaled.set_index(test.index,  inplace=True);

train.index

train_scaled

test_scaled

type(test_scaled.index)

# Save Data and Scaler

import joblib

scaler_filename = "scaler.save"
joblib.dump(scaler, scaler_filename) 

train.to_csv('01train.csv')
test.to_csv('01test.csv')

train_scaled.to_csv('01train_scaled.csv')
test_scaled.to_csv('01test_scaled.csv')
