==> Oakland/03d_DPP_PreviousWeek.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import joblib

TrainVal_scaled = pd.read_csv('01train_scaled.csv', index_col=0, parse_dates=True)
Test_scaled     = pd.read_csv('01test_scaled.csv', index_col=0, parse_dates=True)
Ext = TrainVal_scaled[:4320]

TrainVal_scaled

Test_scaled

### Total Data

df = pd.concat([TrainVal_scaled, Test_scaled])

df

df.drop(columns = ['f1', 'f2', 'f3', 'f5', 'f6'], inplace=True);

df

df.iloc[4320]          # First sample

df.iloc[2304]          # -1 week (-2016 samples)

df.iloc[288]          # -2 week (-4032 samples)

# Generate Weekly Samples

a = np.zeros((2, 7))
b = np.zeros((2, 7))
x = 4320                           # Start

for i in range(len(df.iloc[x:])):
    for j in range(2):
        for k in range(7):
            if j == 0:
                a[j][k] = df.iloc[x-2019+i+k]
            else:
                a[j][k] = df.iloc[x-4035+i+k]
    b = np.append(b, a, axis=0)

b = np.delete(b, np.s_[0:2], 0).reshape(-1, 2, 7)

b[0]

b[1]

b[0].flatten()

a = b.flatten().reshape(-1,14)

a[0]

len(a) == len(TrainVal_scaled) + len(Test_scaled) - len(Ext)

### Split into corresponding test and train

a.shape

Ext.shape[0]

TrainVal_scaled[4320:].shape

train = a[:TrainVal_scaled[4320:].shape[0]]

train.shape

test = a[TrainVal_scaled[4320:].shape[0]:]

test.shape

Test_scaled.shape

#### Confirmation

test[0]

TrainVal_scaled.iloc[-2016]['f4']  

# Save Data

joblib.dump(train, "002weeks_train.save");
joblib.dump(test, "002weeks_test.save");
