==> Oakland/05_Sampe Generators_II_Wide-Multi.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import joblib

# General Period
train_period = joblib.load("00Periodicity_train.save") 
test_period = joblib.load("00Periodicity_test.save") 

# 2 Week priodicity
train_2week = joblib.load("002weeks_train.save") 
test_2week = joblib.load("002weeks_test.save") 

train_period.shape

train_2week.shape

test_period.shape

test_2week.shape

train_period[0]

train_2week[0]

np.concatenate((train_period[0], train_2week[0]), axis=0)

### Train

b = np.zeros(63)

for i in range(len(train_period)):
    a = np.concatenate((train_period[i], train_2week[i]), axis=0)
    b = np.append(b, a, axis=0)

b = b.reshape(-1,63)
Xtrain = np.delete(b, np.s_[0], 0)

Xtrain[0] == np.concatenate((train_period[0], train_2week[0]), axis=0)

Xtrain.shape

### Test

b = np.zeros(63)

for i in range(len(test_period)):
    a = np.concatenate((test_period[i], test_2week[i]), axis=0)
    b = np.append(b, a, axis=0)
    
b = b.reshape(-1,63)
Xtest = np.delete(b, np.s_[0], 0)

Xtest[0] == np.concatenate((test_period[0], test_2week[0]), axis=0)

Xtest.shape

# Samples

### 5 mins

# Delete first 15 samples
a = np.delete(Xtrain, np.s_[0:15], 0)
b = np.delete(Xtest, np.s_[0:15], 0)

len(a)

len(b)

# Save 

np.save("Wide_train_5", a)
np.save("Wide_test_5", b)

### 15 mins

# Delete first 17 samples
a = np.delete(Xtrain, np.s_[0:17], 0)
b = np.delete(Xtest, np.s_[0:17], 0)

len(a)

len(b)

# Save 

np.save("Wide_train_15", a)
np.save("Wide_test_15", b)

### 30 mins

# Delete first 20 samples
a = np.delete(Xtrain, np.s_[0:20], 0)
b = np.delete(Xtest, np.s_[0:20], 0)

len(a)

len(b)

# Save 

np.save("Wide_train_30", a)
np.save("Wide_test_30", b)

### 60 mins

# Delete first 27 samples
a = np.delete(Xtrain, np.s_[0:26], 0)
b = np.delete(Xtest, np.s_[0:26], 0)

len(a)

len(b)

# Save 

np.save("Wide_train_60", a)
np.save("Wide_test_60", b)
