==> Oakland/05a_Sample Generators_I_Deep-Test-July.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import joblib

#### Read training data from csv files

Test      = pd.DataFrame(joblib.load("0000Test.save")) 

columns = ['f1', 'f2', 'f3', 'f4', 'f5', 'f6']
Test.columns = columns

date_range = pd.date_range('2018-07-01 00:00', '2018-07-29 23:59', freq='5Min')

test = Test[4320:].copy().reset_index(drop=True)

Test['Timestamp'] = date_range

# 5-minute (1 step ahead) foreast

length = 15
step = 1
k = -1

a = np.zeros((15, 6))
b = np.zeros((1,))
x = np.zeros((15, 6))
y = np.zeros((1,))


for i in range(length, (len(test)-step+1)):
    b[0] = test["f4"][i+step-1]
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)

x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 6)
y = np.delete(y, np.s_[0], 0)

x.shape

y.shape

# Checked as correct

# Save 

np.savez_compressed("Deep_July_test_5", x=x , y=y)

# 15-minute (3 step ahead) forecast

length = 15
step = 3
k = -1

a = np.zeros((15, 6))
b = np.zeros((1,))
x = np.zeros((15, 6))
y = np.zeros((1,))


for i in range(length, (len(test)-step+1)):
    b[0] = test["f4"][i+step-1]
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 6)
y = np.delete(y, np.s_[0], 0)

x.shape

y.shape

# Checked as correct

# Save 

np.savez_compressed("Deep_July_test_15", x=x , y=y)

# 30-minute (6 step ahead) forecast

length = 15
step = 6
k = -1

a = np.zeros((15, 6))
b = np.zeros((1,))
x = np.zeros((15, 6))
y = np.zeros((1,))


for i in range(length, (len(test)-step+1)):
    b[0] = test["f4"][i+step-1]
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 6)
y = np.delete(y, np.s_[0], 0)

#x = np.delete(x, np.s_[0:4300], 0)
#y = np.delete(y, np.s_[0:4300], 0)

x.shape

y.shape

# Checked as correct

# Save 

np.savez_compressed("Deep_July_test_30", x=x , y=y)

# 60-minute (12 step ahead) forecast

length = 15
step = 12
k = -1

a = np.zeros((15, 6))
b = np.zeros((1,))
x = np.zeros((15, 6))
y = np.zeros((1,))


for i in range(length, (len(test)-step+1)):
    b[0] = test["f4"][i+step-1]
    k = k+1
    for j in range(length):
        if j+k <= (len(test) + length - step):
            a[j] = test.iloc[j+k]
            
    x = np.append(x, a, axis=0)
    y = np.append(y, b, axis=0)
    
x = np.delete(x, np.s_[0:15], 0).reshape(-1, 15, 6)
y = np.delete(y, np.s_[0], 0)

#x = np.delete(x, np.s_[0:4294], 0)
#y = np.delete(y, np.s_[0:4294], 0)

x.shape

y.shape

# Checked as correct

# Save 

np.savez_compressed("Deep_July_test_60", x=x , y=y)
